+++ Assessments.Frontend.Web/Infrastructure/AlienSpecies/FilterItemHelpers.cs
@@ -1,11 +1,59 @@
-﻿namespace Assessments.Frontend.Web.Infrastructure
+﻿using System;
+using System.Linq;
+using Assessments.Mapping.AlienSpecies;
+using Assessments.Shared.Helpers;
+
+namespace Assessments.Frontend.Web.Infrastructure.AlienSpecies
 {
+    public class Areas
+    {
+        public static readonly Filter.FilterItem[] AlienSpecies2023Areas = Enum.GetValues<AlienSpeciesAssessment2023EvaluationContext>()
+            .Select(x => new Filter.FilterItem
+            {
+                Name = x.DisplayName(),
+                NameShort = x.ToString()
+            }).ToArray();
+    }
+
+    public class Categories
+    {
+        public static readonly Filter.FilterItem[] AlienSpecies2023Categories = Enum.GetValues<AlienSpeciesAssessment2023Category>()
+            .Select(x => new Filter.FilterItem
+            {
+                NameShort = x.ToString(),
+                Name = x.DisplayName().ToLowerInvariant(),
+                Description = x.DisplayName()
+            }).ToArray();
+    }
+
+    public class TaxonRank
+    {
+        public static readonly Filter.FilterItem[] AlienSpecies2023TaxonRanks =
+        {
+            new()
+            {
+                Name = "Art",
+                NameShort = "a"
+            },
+            new()
+            {
+                Name = "Underart",
+                NameShort = "ua"
+            },
+            new()
+            {
+                Name = "Varietet",
+                NameShort = "v"
+            }
+        };
+    }
+
     public class SpeciesGroups
     {
         // All NameShort for speciesGroup starts with 's' to not confuse them with other short names. 
         private static readonly Filter.FilterItem[] AlienSpecies2023Algae =
         {
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Brunagler",
                 NameShort = "sbr",
@@ -13,7 +61,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Alger.svg",
                 Description = "Phaeophyceae"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Grønnalger",
                 NameShort = "sga",
@@ -21,7 +69,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Alger.svg",
                 Description = "Chlorophyta"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Rødalger",
                 NameShort = "sra",
@@ -33,7 +81,7 @@
 
         private static readonly Filter.FilterItem[] AlienSpecies2023Insects =
         {
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Biller",
                 NameShort = "sbi",
@@ -41,7 +89,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Biller.svg",
                 Description = "Coleoptera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Børstehaler",
                 NameShort = "sbo",
@@ -49,7 +97,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/B%c3%b8rstehale.svg",
                 Description = "Zygentoma"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Knelere",
                 NameShort = "skn",
@@ -57,7 +105,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Knelere.svg",
                 Description = "Mantodea"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Lus og lopper",
                 NameShort = "sll",
@@ -65,7 +113,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/LusogLopper.svg",
                 Description = "Phthiraptera og Siphonatera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Nebbmunner",
                 NameShort = "sne",
@@ -73,7 +121,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Nebbmunner.svg",
                 Description = "Hemiptera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Rettvinger",
                 NameShort = "srv",
@@ -81,7 +129,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Rettvinger.svg",
                 Description = "Orthoptera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Sommerfugler",
                 NameShort = "ssf",
@@ -89,7 +137,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Sommerfugler.svg",
                 Description = "Lepidoptera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Støvlus",
                 NameShort = "ssl",
@@ -97,7 +145,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/St%c3%b8vlus.svg",
                 Description = "Psocoptera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Tovinger",
                 NameShort = "sto",
@@ -105,7 +153,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Tovinger.svg",
                 Description = "Diptera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Trips",
                 NameShort = "str",
@@ -113,7 +161,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Trips.svg",
                 Description = "Thysanoptera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Vepser",
                 NameShort = "sve",
@@ -125,7 +173,7 @@
 
         public static readonly Filter.FilterItem[] AlienSpecies2023Crustacean =
         {
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Storkrepser",
                 NameShort = "sst",
@@ -133,7 +181,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Storkrepser.svg",
                 Description = "Malacostraca"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Bladfotinger",
                 NameShort = "sbf",
@@ -141,7 +189,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Bladfotinger.svg",
                 Description = "Branchiopoda"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Maxillopoda",
                 NameShort = "sma",
@@ -153,7 +201,7 @@
 
         public static readonly Filter.FilterItem[] AlienSpecies2023SpeciesGroups =
         {
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Alger",
                 NameShort = "sal",
@@ -162,7 +210,7 @@
                 Description = "Rhodophyta, Chlorophyta og Phaeophyceae",
                 SubGroup = AlienSpecies2023Algae
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Amfibier",
                 NameShort = "sam",
@@ -170,7 +218,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Amfibier.svg",
                 Description = "Amphibia"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Bakterier",
                 NameShort = "sba",
@@ -178,7 +226,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Bakterier.svg",
                 Description = "Bacteria"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Bløtdyr",
                 NameShort = "sbl",
@@ -186,7 +234,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Bl%c3%b8tdyr.svg",
                 Description = "Mollusca"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Edderkoppdyr",
                 NameShort = "sed",
@@ -194,7 +242,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Edderkoppdyr.svg",
                 Description = "Arachnida"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Eggsporesopper",
                 NameShort = "seg",
@@ -202,7 +250,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Eggsporesopper.svg",
                 Description = "Oomycota"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Fisker",
                 NameShort = "sfi",
@@ -210,7 +258,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Fisker.svg",
                 Description = "Actinopterygii"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Flatormer",
                 NameShort = "sfl",
@@ -218,7 +266,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Flatorm.svg",
                 Description = "Actinopterygii"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Fugler",
                 NameShort = "sfu",
@@ -226,7 +274,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Fugler.svg",
                 Description = "Aves"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Havedderkopper",
                 NameShort = "sha",
@@ -234,7 +282,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Havedderkopper.svg",
                 Description = "Pycnogonida"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Hjuldyr",
                 NameShort = "shj",
@@ -242,7 +290,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Hjuldyr.svg",
                 Description = "Rotifera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Insekter",
                 NameShort = "sin",
@@ -251,7 +299,7 @@
                 Description = "Insecta",
                 SubGroup = AlienSpecies2023Insects
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Kammaneter",
                 NameShort = "ska",
@@ -259,7 +307,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Kammaneter.svg",
                 Description = "Ctenophora"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Kappedyr",
                 NameShort = "skd",
@@ -267,7 +315,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Kammaneter.svg",
                 Description = "Ascidiacea (Tunicata)"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Karplanter",
                 NameShort = "skp",
@@ -275,7 +323,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Kammaneter.svg",
                 Description = "Magnoliophyta, Pinophyta og Pteridophyta"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Krepsdyr",
                 NameShort = "skr",
@@ -284,7 +332,7 @@
                 Description = "Crustacea",
                 SubGroup = AlienSpecies2023Crustacean
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Leddormer",
                 NameShort = "sle",
@@ -292,7 +340,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Leddormer.svg",
                 Description = "Annelida"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Mangefotinger",
                 NameShort = "smf",
@@ -300,7 +348,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Mangefotinger.svg",
                 Description = "Myriapoda [Chilopoda og Diplopoda]"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Mosdyr og begerormer",
                 NameShort = "smb",
@@ -308,7 +356,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Mosdyr.svg",
                 Description = "Ectoprocta og Entoprocta"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Moser",
                 NameShort = "smo",
@@ -316,7 +364,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Moser.svg",
                 Description = "Bryophyta og Marchantiophyta"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Nesledyr",
                 NameShort = "snd",
@@ -324,7 +372,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Nesledyr.svg",
                 Description = "Cnidaria"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Pigghuder",
                 NameShort = "spi",
@@ -332,7 +380,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Pigghuder.svg",
                 Description = "Echinodermata"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Reptiler",
                 NameShort = "sre",
@@ -340,7 +388,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Reptiler.svg",
                 Description = "Reptilia"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Rundormer",
                 NameShort = "sru",
@@ -348,7 +396,7 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Rundormer.svg",
                 Description = "Nematoda"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Svamper",
                 NameShort = "ssv",
@@ -356,14 +404,264 @@
                 ImageUrl = "https://design.artsdatabanken.no/icons/Svamper.svg",
                 Description = "Porifera"
             },
-            new Filter.FilterItem()
+            new()
             {
                 Name = "Sopper",
                 NameShort = "sso",
                 InfoUrl = "https://artsdatabanken.no/fremmedartsliste2023/Artsgruppene/sopper",
                 ImageUrl = "https://design.artsdatabanken.no/icons/Sopper.svg",
                 Description = "Fungi"
+            }
+        };
+    }
+
+    public class Regions
+    {
+        public static readonly Filter.FilterItem[] AlienSpecies2023Regions =
+        {
+            new()
+            {
+                Name = "Aust-Agder",
+                NameShort = "AA"
             },
+            new()
+            {
+                Name = "Buskerud",
+                NameShort = "BU"
+            },
+            new()
+            {
+                Name = "Finnmark",
+                NameShort = "FI"
+            },
+            new()
+            {
+                Name = "Hedmark",
+                NameShort = "HE"
+            },
+            new()
+            {
+                Name = "Hordaland",
+                NameShort = "HO"
+            },
+            new()
+            {
+                Name = "Jan Mayen",
+                NameShort = "JM"
+            },
+            new()
+            {
+                Name = "Nordland",
+                NameShort = "NO"
+            },
+            new()
+            {
+                Name = "Nord-Trøndelag",
+                NameShort = "NT"
+            },
+            new()
+            {
+                Name = "Oppland",
+                NameShort = "OP"
+            },
+            new()
+            {
+                Name = "Oslo og Akershus",
+                NameShort = "OA"
+            },
+            new()
+            {
+                Name = "Rogaland",
+                NameShort = "RO"
+            },
+            new()
+            {
+                Name = "Sogn og Fjordane",
+                NameShort = "SF"
+            },
+            new()
+            {
+                Name = "Svalbard med sjøområder",
+                NameShort = "SS"
+            },
+            new()
+            {
+                Name = "Sør-Trøndelag",
+                NameShort = "ST"
+            },
+            new()
+            {
+                Name = "Telemark",
+                NameShort = "TE"
+            },
+            new()
+            {
+                Name = "Troms",
+                NameShort = "TR"
+            },
+            new()
+            {
+                Name = "Vest-Agder",
+                NameShort = "VA"
+            },
+            new()
+            {
+                Name = "Vestfold",
+                NameShort = "VF"
+            },
+            new()
+            {
+                Name = "ØstFold",
+                NameShort = "ØF"
+            },
+            new()
+            {
+                Name = "Barentshavet nord og Polhavet",
+                NameShort = "BP"
+            },
+            new()
+            {
+                Name = "Barentshavet sør",
+                NameShort = "BS"
+            },
+            new()
+            {
+                Name = "Grønlandshavet",
+                NameShort = "GH"
+            },
+            new()
+            {
+                Name = "Nordsjøen og Skagerrak",
+                NameShort = "NS"
+            },
+            new()
+            {
+                Name = "Norskehavet",
+                NameShort = "NH"
+            },
+        };
+
+        public static readonly Filter.FilterItem[] AlienSpecies2023WaterRegions =
+        {
+            new()
+            {
+                Name = "Agder",
+                NameShort = "WAG"
+            },
+            new()
+            {
+                Name = "Bottenhavet",
+                NameShort = "WBH"
+            },
+            new()
+            {
+                Name = "Bottenviken",
+                NameShort = "WBV"
+            },
+            new()
+            {
+                Name = "Innlandet og Viken",
+                NameShort = "WIV"
+            },
+            new()
+            {
+                Name = "Kemijoki",
+                NameShort = "WKJ"
+            },
+            new()
+            {
+                Name = "Møre og Romsdal",
+                NameShort = "WMR"
+            },
+            new()
+            {
+                Name = "Nordland og Jan Mayen",
+                NameShort = "WNJ"
+            },
+            new()
+            {
+                Name = "Norsk-finsk",
+                NameShort = "WNF"
+            },
+            new()
+            {
+                Name = "Rogaland",
+                NameShort = "WRO"
+            },
+            new()
+            {
+                Name = "Torneå",
+                NameShort = "WTO"
+            },
+            new()
+            {
+                Name = "Tornionjoki",
+                NameShort = "WTJ"
+            },
+            new()
+            {
+                Name = "Troms og Finnmark",
+                NameShort = "WTF"
+            },
+            new()
+            {
+                Name = "Trøndelag",
+                NameShort = "WTR"
+            },
+            new()
+            {
+                Name = "Vestfold og Telemark",
+                NameShort = "WVT"
+            },
+            new()
+            {
+                Name = "Vestland",
+                NameShort = "WVE"
+            },
+            new()
+            {
+                Name = "Västerhavet",
+                NameShort = "WVH"
+            }
+        };
+    }
+
+    public class Habitat
+    {
+        public static readonly Filter.FilterItem[] AlienSpecies2023Habitats =
+        {
+            new()
+            {
+                Name = "Fastmarksskogsmark",
+                NameShort = "Skog",
+                ImageUrl = "https://data.artsdatabanken.no//Natur_i_Norge/Natursystem/Typeinndeling/Fastmarkssystemer/Fastmarksskogsmark/foto_408.jpg",
+                Description = "Skogarealer, med unntak av treplantasjer, semi-naturlig eng som er tresatt og skog som påvirkes av flom.",
+                Mapping = "Ensbetydende med NiN 2.0 hovedtypen <a href=\"https://artsdatabanken.no/Pages/171923/Fastmarksskogsmark\" target=\"_blank\" rel=\"noopener\">T4 Fastmarksskogsmark</a>."
+            },
+            new()
+            {
+                Name = "Ferskvannssystemer",
+                NameShort = "Ferskvann",
+                ImageUrl = "https://data.artsdatabanken.no//Natur_i_Norge/Natursystem/Typeinndeling/Limniske_vannmasser/foto_408.jpg",
+                Description = "Samlebetegnelse på naturtyper i ferskvann.",
+                Mapping = "Omfatter alle typene under NiN 2.0 hovedtypegruppene <a href=\"https://artsdatabanken.no/Pages/172024/Ferskvannsbunnsystemer\" target=\"_blank\" rel=\"noopener\">L Ferskvannsbunnsystemer</a> og <a href= \"https://artsdatabanken.no/Pages/172053/Limniske_vannmasser\" target =\"_blank\" rel=\"noopener\">F Limniske vannmasser</a>."
+            },
+            new()
+            {
+                Name = "Semi-naturlig fastmark",
+                NameShort = "Semi-naturlig mark",
+                ImageUrl = "https://data.artsdatabanken.no//Natur_i_Norge/Natursystem/Typeinndeling/Fastmarkssystemer/Kystlynghei/foto_408.jpg",
+                Description = "Samlebetegnelse på åpne eller spredt tresatte fastmarkssystemer som er formet og betinget av tradisjonell hevd.",
+                Mapping = "Omfatter NiN 2.0 hovedtypene <a href=\"https://artsdatabanken.no/Pages/171949/Boreal_hei\" target=\"_blank\" rel=\"noopener\">T31 Boreal hei</a>, <a href=\"//artsdatabanken.no/Pages/171950/Semi-naturlig_eng\" target=\"_blank\" rel=\"noopener\">T32 Semi-naturlig eng</a>, <a href=\"//artsdatabanken.no/Pages/171951/Semi-naturlig_strandeng\" target=\"_blank\" rel=\"noopener\">T33 Semi-naturlig strandeng</a> og <a href=\"//artsdatabanken.no/Pages/171952/Kystlynghei\" target=\"_blank\" rel=\"noopener\">T34 Kystlynghei</a>."
+            },
+            new()
+            {
+                Name = "Våtmarkssystemer",
+                NameShort = "Våtmark",
+                ImageUrl = "https://data.artsdatabanken.no//Natur_i_Norge/Natursystem/Typeinndeling/V%C3%A5tmarkssystemer/foto_408.jpg",
+                Description = "Samlebetegnelse på myr, kilder og andre økosystmer på mer eller mindre vannmettet mark.",
+                Mapping = "Omfatter alle hovedtypene under NiN 2.0 hovedtypegruppen <a href=\"https://artsdatabanken.no/Pages/172028/Vaatmarkssystemer\" target=\"_blank\" rel=\"noopener\">V Våtmarkssystemer</a>."
+            }
         };
     }
 }
