--- Assessments.Frontend.Web/Infrastructure/Regions.cs
@@ -1,216 +1,0 @@
-﻿using DocumentFormat.OpenXml.Drawing.Charts;
-using DocumentFormat.OpenXml.Spreadsheet;
-
-namespace Assessments.Frontend.Web.Infrastructure
-{
-    public class Regions
-    {
-        public static readonly Filter.FilterItem[] AlienSpecies2023Regions =
-        {
-            new Filter.FilterItem()
-            { 
-                Name = "Aust-Agder",
-                NameShort = "AA"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Buskerud",
-                NameShort = "BU"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Finnmark",
-                NameShort = "FI"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Hedmark",
-                NameShort = "HE"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Hordaland",
-                NameShort = "HO"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Jan Mayen",
-                NameShort = "JM"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Nordland",
-                NameShort = "NO"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Nord-Trøndelag",
-                NameShort = "NT"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Oppland",
-                NameShort = "OP"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Oslo og Akershus",
-                NameShort = "OA"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Rogaland",
-                NameShort = "RO"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Sogn og Fjordane",
-                NameShort = "SF"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Svalbard med sjøområder",
-                NameShort = "SS"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Sør-Trøndelag",
-                NameShort = "ST"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Telemark",
-                NameShort = "TE"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Troms",
-                NameShort = "TR"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Vest-Agder",
-                NameShort = "VA"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Vestfold",
-                NameShort = "VF"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "ØstFold",
-                NameShort = "ØF"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Barentshavet nord og Polhavet",
-                NameShort = "BP"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Barentshavet sør",
-                NameShort = "BS"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Grønlandshavet",
-                NameShort = "GH"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Nordsjøen og Skagerrak",
-                NameShort = "NS"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Norskehavet",
-                NameShort = "NH"
-            },
-        };
-
-        public static readonly Filter.FilterItem[] AlienSpecies2023WaterRegions =
-        {
-            new Filter.FilterItem()
-            {
-                Name = "Agder",
-                NameShort = "WAG"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Bottenhavet",
-                NameShort = "WBH"
-            }, 
-            new Filter.FilterItem()
-            {
-                Name = "Bottenviken",
-                NameShort = "WBV"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Innlandet og Viken",
-                NameShort = "WIV"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Kemijoki",
-                NameShort = "WKJ"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Møre og Romsdal",
-                NameShort = "WMR"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Nordland og Jan Mayen",
-                NameShort = "WNJ"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Norsk-finsk",
-                NameShort = "WNF"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Rogaland",
-                NameShort = "WRO"
-            }, 
-            new Filter.FilterItem()
-            {
-                Name = "Torneå",
-                NameShort = "WTO"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Tornionjoki",
-                NameShort = "WTJ"
-            }, 
-            new Filter.FilterItem()
-            {
-                Name = "Troms og Finnmark",
-                NameShort = "WTF"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Trøndelag",
-                NameShort = "WTR"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Vestfold og Telemark",
-                NameShort = "WVT"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Vestland",
-                NameShort = "WVE"
-            }, 
-            new Filter.FilterItem()
-            {
-                Name = "Västerhavet",
-                NameShort = "WVH"
-            },
-        };
-    }
-}
