--- Assessments.Frontend.Web/Infrastructure/TaxonRank.cs
@@ -1,24 +1,0 @@
-﻿namespace Assessments.Frontend.Web.Infrastructure
-{
-    public class TaxonRank
-    {
-        public static Filter.FilterItem[] AlienSpecies2023TaxonRanks =
-        {
-            new Filter.FilterItem()
-            {
-                Name = "Art",
-                NameShort = "a"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Underart",
-                NameShort = "ua"
-            },
-            new Filter.FilterItem()
-            {
-                Name = "Varietet",
-                NameShort = "v"
-            }
-        };
-    }
-}
