+++ Assessments.Frontend.Web/Views/AlienSpecies/2023/ListPartials/_Filters.cshtml
@@ -3,6 +3,7 @@
 @using static Assessments.Frontend.Web.Infrastructure.Helpers;
 @using static Assessments.Frontend.Web.Infrastructure.AlienSpecies.FilterHelpers;
 @using System.Linq;
+@using Assessments.Frontend.Web.Infrastructure.AlienSpecies
 
 @{
     Filter.FilterItem[] AllAreas = Areas.AlienSpecies2023Areas;
