+++ Assessments.Frontend.Web/Views/AlienSpecies/2023/ListPartials/_ListView.cshtml
@@ -1,5 +1,7 @@
 @model AlienSpeciesListViewModel
 @using static Assessments.Frontend.Web.Infrastructure.Helpers
+@using Assessments.Frontend.Web.Infrastructure.AlienSpecies
+@using Microsoft.AspNetCore.Mvc.TagHelpers
 
 @{
     Filter.FilterItem[] speciesgroup = SpeciesGroups.AlienSpecies2023SpeciesGroups;
